src/requirements.txt
argparse~=1.4.0
ffmpeg-python~=0.2.0
telethon~=1.42.0
distro~=1.9.0
curl_cffi~=0.12.0
requests~=2.32.0
websocket-client~=1.6.3
